# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Profile {
  name: string;
  designation: string;
  department: string;
  role: string;
  assignment: string;
  education: string;
  experience: number;
  futureGoal: string;
  avatarInitials: string;
}

export interface AuthSession {
  token: string;
  user: Profile;
}

export interface SkillScore {
  name: string;
  category: string;
  score: number;
  target: number;
  delta: number;
  level: string;
}

export interface DomainScore {
  name: string;
  score: number;
  color: string;
}

export interface Competency {
  overall: number;
  level: string;
  domains: DomainScore[];
  skills: SkillScore[];
  strongSkills: string[];
  gaps: string[];
  criticalGap: string;
  insight: string;
}

export interface Recommendation {
  id: string;
  title: string;
  provider: string;
  skill: string;
  level: string;
  targetLevel: string;
  duration: string;
  priority: string;
  reason: string;
  kind: string;
}

export interface Course {
  id: string;
  title: string;
  provider: string;
  skill: string;
  duration: string;
  progress: number;
  lessons: number;
  totalLessons: number;
  color: string;
  description: string;
}

export interface CourseModule {
  title: string;
  duration: string;
  completed: boolean;
}

export type CourseDetail = Course & {
  objectives: string[];
  modules: CourseModule[];
};

export interface IntegrationCourse {
  title: string;
  status: string;
  provider: string;
  relevance: number;
}

export interface TrainingProgram {
  title: string;
  dates: string;
  mode: string;
  duration: string;
  relevance: number;
}

export interface IntegrationHub {
  igot: IntegrationCourse[];
  nssta: TrainingProgram[];
}

export interface GardenPlant {
  skill: string;
  type: string;
  growth: number;
  score: number;
  status: string;
  color: string;
  x: number;
  y: number;
}

export interface Garden {
  growth: number;
  totalHours: number;
  plantsUnlocked: number;
  skillsCultivated: number;
  streak: number;
  recentlyGrown: string;
  plants: GardenPlant[];
}

export interface SkillGene {
  skill: string;
  category: string;
  score: number;
  level: string;
  color: string;
}

export interface SkillDna {
  diversityScore: number;
  strongestGene: string;
  fastestGrowing: string;
  recentlyAdded: string;
  genes: SkillGene[];
}

export interface Shield {
  name: string;
  tier: string;
  unlocked: boolean;
  progress: number;
  requirement: string;
}

export interface Achievements {
  current: string;
  currentTier: string;
  next: string;
  nextProgress: number;
  shields: Shield[];
}

export interface ActivityItem {
  title: string;
  subtitle: string;
  time: string;
  type: string;
}

export interface Dashboard {
  profile: Profile;
  competency: Competency;
  recommendations: Recommendation[];
  garden: Garden;
  dna: SkillDna;
  achievements: Achievements;
  courses: Course[];
  activity: ActivityItem[];
  learningHours: number;
  streak: number;
}

export interface DepartmentMetric {
  name: string;
  score: number;
  learners: number;
}

export interface GapMetric {
  skill: string;
  percentage: number;
}

export interface TrendMetric {
  month: string;
  score: number;
}

export interface Analytics {
  totalLearners: number;
  averageCompetency: number;
  completionRate: number;
  learningHours: number;
  departments: DepartmentMetric[];
  gaps: GapMetric[];
  trend: TrendMetric[];
}

export interface AssessmentSubmission {
  answers: number[];
}

export interface AssessmentResult {
  score: number;
  level: string;
  strong: string[];
  improve: string[];
  critical: string;
  message: string;
}

export interface QuizGenerateInput {
  title: string;
  content: string;
}

export interface QuizQuestion {
  id: number;
  question: string;
  options: string[];
  answer: number;
  explanation: string;
  difficulty: string;
}

export interface Quiz {
  id: string;
  title: string;
  topic: string;
  totalQuestions: number;
  questions: QuizQuestion[];
}

export interface QuizSubmission {
  answers: number[];
}

export interface QuizResult {
  score: number;
  correct: number;
  total: number;
  feedback: string;
  skillUpdate: string;
}

export interface LearningTimeInput {
  courseId: string;
  minutes: number;
}

export interface ProgressUpdate {
  learningHours: number;
  gardenGrowth: number;
  shieldProgress: number;
  message: string;
}