# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentMetric } from './departmentMetric';
import type { GapMetric } from './gapMetric';
import type { TrendMetric } from './trendMetric';

export interface Analytics {
  totalLearners: number;
  averageCompetency: number;
  completionRate: number;
  learningHours: number;
  departments: DepartmentMetric[];
  gaps: GapMetric[];
  trend: TrendMetric[];
}