# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: string;
  title: string;
  provider: string;
  skill: string;
  duration: string;
  progress: number;
  lessons: number;
  totalLessons: number;
  color: string;
  description: string;
}