# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Achievements } from './achievements';
import type { ActivityItem } from './activityItem';
import type { Competency } from './competency';
import type { Course } from './course';
import type { Garden } from './garden';
import type { Profile } from './profile';
import type { Recommendation } from './recommendation';
import type { SkillDna } from './skillDna';

export interface Dashboard {
  profile: Profile;
  competency: Competency;
  recommendations: Recommendation[];
  garden: Garden;
  dna: SkillDna;
  achievements: Achievements;
  courses: Course[];
  activity: ActivityItem[];
  learningHours: number;
  streak: number;
}