# lib/api-zod/src/generated/types/domainScore.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DomainScore {
  name: string;
  score: number;
  color: string;
}