# lib/api-zod/src/generated/types/garden.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GardenPlant } from './gardenPlant';

export interface Garden {
  growth: number;
  totalHours: number;
  plantsUnlocked: number;
  skillsCultivated: number;
  streak: number;
  recentlyGrown: string;
  plants: GardenPlant[];
}