# lib/api-zod/src/generated/types/integrationCourse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IntegrationCourse {
  title: string;
  status: string;
  provider: string;
  relevance: number;
}