# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  name: string;
  designation: string;
  department: string;
  role: string;
  assignment: string;
  education: string;
  experience: number;
  futureGoal: string;
  avatarInitials: string;
}