# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';

export interface Quiz {
  id: string;
  title: string;
  topic: string;
  totalQuestions: number;
  questions: QuizQuestion[];
}