# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  id: string;
  title: string;
  provider: string;
  skill: string;
  level: string;
  targetLevel: string;
  duration: string;
  priority: string;
  reason: string;
  kind: string;
}