# lib/api-zod/src/generated/types/shield.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Shield {
  name: string;
  tier: string;
  unlocked: boolean;
  progress: number;
  requirement: string;
}