# lib/api-zod/src/generated/types/skillDna.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillGene } from './skillGene';

export interface SkillDna {
  diversityScore: number;
  strongestGene: string;
  fastestGrowing: string;
  recentlyAdded: string;
  genes: SkillGene[];
}