# lib/api-zod/src/generated/types/skillGene.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillGene {
  skill: string;
  category: string;
  score: number;
  level: string;
  color: string;
}